+++ Contoso.Test.Flow.Test/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace Contoso.Test.Flow.Test
+{
+    internal struct Constants
+    {
+        public const string ProjectFileFullPath = @"C:\LogicBuilder\FlowProjects\Contoso.Test\Contoso.Test.lbproj";
+    }
+}