+++ Contoso.Test.Flow.Test/LambdaExpressions/SelectorTests.cs
@@ -1,4 +1,6 @@
-﻿using AutoMapper;
+﻿using ABIS.LogicBuilder.FlowBuilder.ServiceInterfaces.Configuration;
+using ABIS.LogicBuilder.FlowBuilder.ServiceInterfaces.Configuration.Initialization;
+using AutoMapper;
 using AutoMapper.Extensions.ExpressionMapping;
 using Contoso.BSL.AutoMapperProfiles;
 using Contoso.Contexts;
@@ -24,7 +26,7 @@ using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Contoso.Test.Flow.Test.LmbdaExpressions
+namespace Contoso.Test.Flow.Test.LambdaExpressions
 {
     [Collection("DatabaseCollection")]
     public class SelectorTests
@@ -52,6 +54,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_group_students_by_enrollment_date()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -85,6 +99,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_select_course_credits_for_grid_filter()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -118,6 +144,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_select_course_credits_text_and_value()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -151,6 +189,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_select_course_id_for_grid_column_filter()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -184,6 +234,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_select_course_title_and_id()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -217,6 +279,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_select_department_name_and_id()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -250,6 +324,18 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         public async Task Selector_select_instructor_full_name_and_id()
         {
             //arrange
+            IConfigurationService _configurationService = serviceProvider.GetRequiredService<IConfigurationService>();
+            IFragmentListInitializer _fragmentListInitializer = serviceProvider.GetRequiredService<IFragmentListInitializer>();
+            IFunctionListInitializer _functionListInitializer = serviceProvider.GetRequiredService<IFunctionListInitializer>();
+            IVariableListInitializer _variableListInitializer = serviceProvider.GetRequiredService<IVariableListInitializer>();
+            ILoadProjectProperties _loadProjectProperties = serviceProvider.GetRequiredService<ILoadProjectProperties>();
+            IConstructorListInitializer _constructorListInitializer = serviceProvider.GetRequiredService<IConstructorListInitializer>();
+            _configurationService.ProjectProperties = _loadProjectProperties.Load(Constants.ProjectFileFullPath);
+            _configurationService.ConstructorList = _constructorListInitializer.InitializeList();
+            _configurationService.FragmentList = _fragmentListInitializer.InitializeList();
+            _configurationService.FunctionList = _functionListInitializer.InitializeList();
+            _configurationService.VariableList = _variableListInitializer.InitializeList();
+
             IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
             IMappingOperations mappingOperations = serviceProvider.GetRequiredService<IMappingOperations>();
             ISchoolRepository schoolRepository = serviceProvider!.GetRequiredService<ISchoolRepository>();
@@ -311,7 +397,7 @@ namespace Contoso.Test.Flow.Test.LmbdaExpressions
         [MemberNotNull(nameof(serviceProvider))]
         private void Initialize()
         {
-            serviceProvider = new ServiceCollection()
+            serviceProvider = ABIS.LogicBuilder.FlowBuilder.Program.ServiceCollection
                 .AddDbContext<SchoolContext>
                 (
                     options => options.UseSqlServer